accounts/requirements.txt
fastapi[all]==0.78.0
uvicorn[standard]==0.17.6
asyncpg==0.27.0
sqlmodel==0.0.8
alembic==1.9.4
python-jose[cryptopgraphy]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.23.3
pytest==7.2.2
pytest-asyncio==0.20.3
aiosqlite==0.18.0
